Move task files
